MNT #594 _utils & utils
