+++ apstools/utils/_core.py
