+++ apstools/utils/device_info.py
