+++ apstools/utils/tests/test_listdevice.py
